For homework #5. Refactoring of boxGrid.
The purpose of this file is to draw a 3x3 grid of boxes. For this commit
I've refactored the code into 2 new operator definitions.

Operator boxLine, this function will draw a row of three boxes from a
provided start coordinate.

Operator boxGrid, this function will draw a grid of boxes from the start
point of the page by calling boxLine 3 times.

			***** TODO *****
* Be able to specify the number boxes that boxLine will generate
* Be able to specify the length of the boxGrid side.
* Rename boxGrid operator to boxSquareGrid